Lab/Data_Lakehouse_lab/requirements.txt
# Data Lakehouse Lab - Integrated Stack

# Apache Spark
pyspark==3.5.0

# Apache Iceberg
pyiceberg==0.6.0

# Apache Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Apache Airflow
apache-airflow==3.1.1
apache-airflow-providers-apache-kafka==4.4.0
apache-airflow-providers-apache-spark==4.5.0
apache-airflow-providers-postgres==5.10.0

# dbt
dbt-core==1.7.0
dbt-postgres==1.7.0

# Great Expectations
great-expectations==0.18.15

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data processing
pandas==2.1.0
numpy==1.24.3
pyarrow==14.0.1

# Jupyter
jupyter==1.0.0
ipykernel==6.25.2
notebook==7.0.6

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

